Remove file comparison utilities
